requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
requests==2.18.4
pillow==8.3.2
tensorflow==1.12.0
torch==1.2.0
torchvision==0.2.0
